+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+## unreleased
+### Changed
+- Replaced Ensembl rest liftover service with liftover API from the Broad Institute (#6293)
+
 ## [4.112]
 ### Added
 - Contigs on exported variants VCF files (managed and causative variants) (#6310)